cuda/requirements.txt
# PyTorch/CUDA requirements for mHC-POC
torch>=2.0.0
pyyaml>=6.0
matplotlib>=3.7.0

# Optional: FlashAttention-2 (requires CUDA)
# flash-attn>=2.0.0
